refactor: improve security, modularity and standards - Secure postMessage, extract Swap components, modernization
